graphql: Add HTTP401-wrapper @auth_required for gql actions
- Create @auth_required wrapper that throws custom
  GraphqlAuthenticationError if user is not authenticated
- Create DRFAuthenticatedGraphQLView which wraps GraphQLView
  and re-raises DRF.AuthenticationFailed if
  GraphqlAuthenticationError is raised. DRF will then return 401
- Create AuthenticatedGraphQLView which wraps GraphQLView
  and returns

Note: DRFAuthenticatedGraphQL view This is a custom workaround.
It raises an HTTP 401 error, which is not a part of the GraphQL
spec. It may fit some frontend apps for a transition period to
GraphQL, but should not be seen as a permanent solution.
AuthenticatedGraphQLView is more in line with GraphQL spec, and
shoul be preferred if applicable for you.